Tirando a declaração dos modelos de models/__init__.py
