+++ indexer/mesos/models/taskadded.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+from indexer.mesos.models.spec import TaskSpec
+
+
+class MesosTaskAddedEvent(BaseModel):
+    task: TaskSpec
